requirements.txt
youtube-transcript-api
google-genai
python-dotenv
pytubefix